Fix cs SSH API objects leak by Cancellation callback
SSH APIs may be called with a long-living `CancellationToken`. The API
must not register cancellation callback without unregistering it.

Replace cancellation callback registration with `Task.WaitAsync()` in
some places.

Fix `TaskExtensions` and add unit tests for it. The issue was that
`Task.WaitAny` never throws and just returns the first to complete task.
If that task was due to cancellation, it would not observe this, and,
instead, synchronously wait for the main `task` to finish.

I verified that memory leak of `SshSession` objects pinned by
`CancellationToken` arg has been fixed in local test Dev Tunnels service
using the SSH library with the fix.

The bug and the fix is specific to cs.