Fix TaskChain scheduling being not sequential (#102)
Fix for https://github.com/devdiv-microsoft/basis-planning/issues/1618

The problem was that `TaskChain` scheduled continuations to keep the task order for task-producing functions, not for the task that they produce. The continuation task was `Task<Task>` The fix was to use `Unwrap` so the continuation task waits for the async code and its result is just `Task`

Add unit tests for more overlapping session requests.
Add unit tests for `TaskChain` class.